refactor(global): Constants 클래스 final 처리 및 네이밍 개선
