+++ src/main/java/blackjack/global/ParticipantConstants.java
@@ -1,10 +1,10 @@
 package blackjack.global;
 
-public class Constants {
+public final class ParticipantConstants {
 
     public static final int DEALER_HIT_THRESHOLD = 16;
 
-    private Constants() {
+    private ParticipantConstants() {
 
     }
