+++ src/main/java/blackjack/view/OutputView.java
@@ -1,6 +1,6 @@
 package blackjack.view;
 
-import static blackjack.global.Constants.DEALER_HIT_THRESHOLD;
+import static blackjack.global.ParticipantConstants.DEALER_HIT_THRESHOLD;
 
 import blackjack.domain.Dealer;
 import blackjack.domain.Participant;
